requirements.txt
langchain_chroma>=0.1.4
langchain_community>=0.3.10
langchain_core>=0.3.22
langchain_openai>=0.2.11
langchain_huggingface>=0.0.6
langchain_text_splitters>=0.3.2
openai>=1.55.3
pydub>=0.25.1
PyPDF>=3.1.0
python-dotenv==1.0.1
gTTS
langchain-ollama
PyPDF2
torch
torchaudio
transformers
cuda-python
nvidia-cuda-runtime-cu12
huggingface_hub>=0.20.1
sentence_transformers>=2.2.2
chromadb>=0.4.18
google-generativeai
langchain-google-genai
typing
fastapi
uvicorn
python-multipart
edge-tts